Extract polyhedron tests
